refactor: add Loading interface to handle loading state in the application
