+++ src/constants/index.ts
@@ -1,4 +1,3 @@
 export const { VITE_BASE_URL } = import.meta.env;
 
 export * from './enums';
-export * from './types';
