+++ src/interfaces/loading.ts
@@ -1,0 +1,6 @@
+import { Status } from '@/constants';
+
+export interface Loading {
+  status: Status;
+  message: string;
+}
